Split PropertyMetadataMapper out of old ContentTypes
