+++ src/Sulu/Bundle/AdminBundle/Tests/Unit/Metadata/FormMetadata/Loader/FormXmlLoaderTest.php
@@ -22,10 +22,9 @@ use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Parser\PropertiesXmlParser;
 use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Parser\SchemaXmlParser;
 use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\Parser\TagXmlParser;
 use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\SchemaMetadataProvider;
+use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\PropertyMetadataMapper\BlockPropertyMetadataMapper;
 use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\PropertyMetadataMapperRegistry;
 use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\SchemaMetadata;
-use Sulu\Component\Content\ContentTypeManagerInterface;
-use Sulu\Component\Content\Types\BlockContentType;
 use Symfony\Component\DependencyInjection\Container;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
@@ -60,10 +59,8 @@ class FormXmlLoaderTest extends TestCase
         $container = new Container();
         $propertyMetadataMapperRegistry = new PropertyMetadataMapperRegistry($container);
         $schemaMetadataProvider = new SchemaMetadataProvider($propertyMetadataMapperRegistry);
-        $blockMetadataProvider = new BlockContentType(
-            $this->prophesize(ContentTypeManagerInterface::class)->reveal(),
+        $blockMetadataProvider = new BlockPropertyMetadataMapper(
             $schemaMetadataProvider,
-            [],
         );
         $container->set('block', $blockMetadataProvider);
