+++ src/Sulu/Bundle/RouteBundle/Content/Type/PageTreeRouteContentType.php
@@ -15,13 +15,6 @@ use Doctrine\ORM\EntityManagerInterface;
 use PHPCR\ItemNotFoundException;
 use PHPCR\NodeInterface;
 use PHPCR\PropertyType;
-use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\FieldMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\AnyOfsMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\NullMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\ObjectMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\PropertyMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\PropertyMetadataMapperInterface;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\StringMetadata;
 use Sulu\Bundle\RouteBundle\Entity\Route;
 use Sulu\Bundle\RouteBundle\Entity\RouteRepositoryInterface;
 use Sulu\Bundle\RouteBundle\Generator\ChainRouteGeneratorInterface;
@@ -34,7 +27,7 @@ use Sulu\Component\DocumentManager\DocumentRegistry;
 /**
  * Provides page_tree_route content-type.
  */
-class PageTreeRouteContentType extends SimpleContentType implements PropertyMetadataMapperInterface
+class PageTreeRouteContentType extends SimpleContentType
 {
     public const NAME = 'page_tree_route';
 
@@ -225,37 +218,5 @@ class PageTreeRouteContentType extends SimpleContentType implements PropertyMeta
     private function getRouteRepository(): RouteRepositoryInterface
     {
         return $this->entityManager->getRepository(Route::class);
-    }
-
-    public function mapPropertyMetadata(FieldMetadata $fieldMetadata): PropertyMetadata
-    {
-        $mandatory = $fieldMetadata->isRequired();
-
-        $pageMetadata = new ObjectMetadata([
-            new PropertyMetadata('uuid', $mandatory, new StringMetadata()),
-            new PropertyMetadata('path', $mandatory, new StringMetadata()),
-        ]);
-
-        if (!$mandatory) {
-            $pageMetadata = new AnyOfsMetadata([
-                new NullMetadata(),
-                $pageMetadata,
-            ]);
-        }
-
-        $pageTreeRouteMetadata = new ObjectMetadata([
-            new PropertyMetadata('page', $mandatory, $pageMetadata),
-            new PropertyMetadata('path', false, new StringMetadata()),
-            new PropertyMetadata('suffix', false, new StringMetadata()),
-        ]);
-
-        if (!$mandatory) {
-            $pageTreeRouteMetadata = new AnyOfsMetadata([
-                new NullMetadata(),
-                $pageTreeRouteMetadata,
-            ]);
-        }
-
-        return new PropertyMetadata($fieldMetadata->getName(), $mandatory, $pageTreeRouteMetadata);
     }
 }
