+++ src/Sulu/Component/Content/Tests/Unit/Types/NumberTest.php
@@ -18,19 +18,12 @@ use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
-use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\FieldMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\OptionMetadata;
 use Sulu\Component\Content\Compat\PropertyInterface;
 use Sulu\Component\Content\Types\Number;
 
 class NumberTest extends TestCase
 {
     use ProphecyTrait;
-
-    /**
-     * @var string
-     */
-    private $template;
 
     /**
      * @var Number
@@ -58,7 +51,7 @@ class NumberTest extends TestCase
         $this->property = $this->prophesize(PropertyInterface::class);
         $this->nodeProperty = $this->prophesize(NodePropertyInterface::class);
 
-        $this->number = new Number($this->template);
+        $this->number = new Number();
     }
 
     public function testRead(): void
@@ -116,295 +109,5 @@ class NumberTest extends TestCase
         $this->node->hasProperty('i18n:de-test')->willReturn(true)->shouldBeCalled();
         $this->node->getProperty('i18n:de-test')->willReturn($this->nodeProperty->reveal())->shouldBeCalled();
         $this->number->write($this->node->reveal(), $this->property->reveal(), 1, 'sulu_io', 'de', null);
-    }
-
-    /**
-     * @return array<string, mixed>
-     */
-    private function getNullSchema(): array
-    {
-        return [
-            'type' => 'null',
-        ];
-    }
-
-    public function testMapPropertyMetadata(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'anyOf' => [
-                $this->getNullSchema(),
-                [
-                    'type' => 'number',
-                ],
-            ],
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataRequired(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-        $fieldMetadata->setRequired(true);
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'type' => 'number',
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataMultipleOfFloat(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-        $multipleOfOption = new OptionMetadata();
-        $multipleOfOption->setName('multiple_of');
-        $multipleOfOption->setValue('0.5'); // float values are by the XmlParserTrait always strings
-        $fieldMetadata->addOption($multipleOfOption);
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'anyOf' => [
-                $this->getNullSchema(),
-                [
-                    'type' => 'number',
-                    'multipleOf' => 0.5,
-                ],
-            ],
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataMultipleOfIntegerish(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-        $multipleOfOption = new OptionMetadata();
-        $multipleOfOption->setName('multiple_of');
-        $multipleOfOption->setValue('2');
-        $fieldMetadata->addOption($multipleOfOption);
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'anyOf' => [
-                $this->getNullSchema(),
-                [
-                    'type' => 'integer',
-                    'multipleOf' => 2.0,
-                ],
-            ],
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataMultipleOfInteger(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-        $multipleOfOption = new OptionMetadata();
-        $multipleOfOption->setName('multiple_of');
-        $multipleOfOption->setValue(2);
-        $fieldMetadata->addOption($multipleOfOption);
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'anyOf' => [
-                $this->getNullSchema(),
-                [
-                    'type' => 'integer',
-                    'multipleOf' => 2.0,
-                ],
-            ],
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataMultipleOfTooSmall(): void
-    {
-        $this->expectExceptionMessage('Parameter "multiple_of" of property "property-name" needs to be greater than "0"');
-
-        $fieldMetadata = new FieldMetadata('property-name');
-        $multipleOfOption = new OptionMetadata();
-        $multipleOfOption->setName('multiple_of');
-        $multipleOfOption->setValue(0);
-        $fieldMetadata->addOption($multipleOfOption);
-
-        $this->number->mapPropertyMetadata($fieldMetadata);
-    }
-
-    public function testMapPropertyMetadataMinAndMaxMultipleOfInvalidType(): void
-    {
-        $this->expectExceptionMessage('Parameter "multiple_of" of property "property-name" needs to be either null or numeric');
-
-        $fieldMetadata = new FieldMetadata('property-name');
-        $multipleOfOption = new OptionMetadata();
-        $multipleOfOption->setName('multiple_of');
-        $multipleOfOption->setValue('invalid-value');
-        $fieldMetadata->addOption($multipleOfOption);
-
-        $this->number->mapPropertyMetadata($fieldMetadata);
-    }
-
-    public function testMapPropertyMetadataMinAndMax(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-        $minOption = new OptionMetadata();
-        $minOption->setName('min');
-        $minOption->setValue(2);
-        $maxOption = new OptionMetadata();
-        $maxOption->setName('max');
-        $maxOption->setValue(3);
-        $fieldMetadata->addOption($minOption);
-        $fieldMetadata->addOption($maxOption);
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'anyOf' => [
-                $this->getNullSchema(),
-                [
-                    'type' => 'number',
-                    'minimum' => 2,
-                    'maximum' => 3,
-                ],
-            ],
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataMinAndMaxMinOnly(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-        $minOption = new OptionMetadata();
-        $minOption->setName('min');
-        $minOption->setValue(2);
-        $fieldMetadata->addOption($minOption);
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'anyOf' => [
-                $this->getNullSchema(),
-                [
-                    'type' => 'number',
-                    'minimum' => 2,
-                ],
-            ],
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataMinAndMaxMaxOnly(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-        $maxOption = new OptionMetadata();
-        $maxOption->setName('max');
-        $maxOption->setValue(3);
-        $fieldMetadata->addOption($maxOption);
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'anyOf' => [
-                $this->getNullSchema(),
-                [
-                    'type' => 'number',
-                    'maximum' => 3,
-                ],
-            ],
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataMinAndMaxWithIntegerishValues(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-        $minOption = new OptionMetadata();
-        $minOption->setName('min');
-        $minOption->setValue('2');
-        $maxOption = new OptionMetadata();
-        $maxOption->setName('max');
-        $maxOption->setValue('3');
-        $fieldMetadata->addOption($minOption);
-        $fieldMetadata->addOption($maxOption);
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'anyOf' => [
-                $this->getNullSchema(),
-                [
-                    'type' => 'number',
-                    'minimum' => 2,
-                    'maximum' => 3,
-                ],
-            ],
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataMinAndMaxWithFloatValues(): void
-    {
-        $fieldMetadata = new FieldMetadata('property-name');
-        $minOption = new OptionMetadata();
-        $minOption->setName('min');
-        $minOption->setValue('1.2'); // float values are by the XmlParserTrait always strings
-        $maxOption = new OptionMetadata();
-        $maxOption->setName('max');
-        $maxOption->setValue('3.4'); // float values are by the XmlParserTrait always strings
-        $fieldMetadata->addOption($minOption);
-        $fieldMetadata->addOption($maxOption);
-
-        $jsonSchema = $this->number->mapPropertyMetadata($fieldMetadata)->toJsonSchema();
-
-        $this->assertEquals([
-            'anyOf' => [
-                $this->getNullSchema(),
-                [
-                    'type' => 'number',
-                    'minimum' => 1.2,
-                    'maximum' => 3.4,
-                ],
-            ],
-        ], $jsonSchema);
-    }
-
-    public function testMapPropertyMetadataMinAndMaxMinInvalidType(): void
-    {
-        $this->expectExceptionMessage('Parameter "min" of property "property-name" needs to be either null or numeric');
-
-        $fieldMetadata = new FieldMetadata('property-name');
-        $minOption = new OptionMetadata();
-        $minOption->setName('min');
-        $minOption->setValue('invalid-value');
-        $fieldMetadata->addOption($minOption);
-
-        $this->number->mapPropertyMetadata($fieldMetadata);
-    }
-
-    public function testMapPropertyMetadataMinAndMaxMaxInvalidType(): void
-    {
-        $this->expectExceptionMessage('Parameter "max" of property "property-name" needs to be either null or numeric');
-
-        $fieldMetadata = new FieldMetadata('property-name');
-        $maxOption = new OptionMetadata();
-        $maxOption->setName('max');
-        $maxOption->setValue('invalid-value');
-        $fieldMetadata->addOption($maxOption);
-
-        $this->number->mapPropertyMetadata($fieldMetadata);
-    }
-
-    public function testMapPropertyMetadataMinAndMaxMaxLowerThanMin(): void
-    {
-        $this->expectExceptionMessage('Because parameter "min" of property "property-name" has value "2", parameter "max" needs to be greater than or equal "2"');
-
-        $fieldMetadata = new FieldMetadata('property-name');
-        $minOption = new OptionMetadata();
-        $minOption->setName('min');
-        $minOption->setValue(2);
-        $maxOption = new OptionMetadata();
-        $maxOption->setName('max');
-        $maxOption->setValue(1);
-        $fieldMetadata->addOption($minOption);
-        $fieldMetadata->addOption($maxOption);
-
-        $this->number->mapPropertyMetadata($fieldMetadata);
     }
 }
