+++ src/Sulu/Component/Content/Types/BlockContentType.php
@@ -12,17 +12,6 @@
 namespace Sulu\Component\Content\Types;
 
 use PHPCR\NodeInterface;
-use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\FieldMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\SchemaMetadataProvider;
-use Sulu\Bundle\AdminBundle\Metadata\FormMetadata\TagMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\AllOfsMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\ArrayMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\ConstMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\IfThenElseMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\PropertyMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\PropertyMetadataMapperInterface;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\RefSchemaMetadata;
-use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\SchemaMetadata;
 use Sulu\Bundle\ReferenceBundle\Application\Collector\ReferenceCollectorInterface;
 use Sulu\Bundle\ReferenceBundle\Infrastructure\Sulu\ContentType\ReferenceContentTypeInterface;
 use Sulu\Component\Content\Compat\Block\BlockPropertyInterface;
@@ -41,14 +30,13 @@ use Sulu\Component\Content\Types\Block\BlockVisitorInterface;
 /**
  * content type for block.
  */
-class BlockContentType extends ComplexContentType implements ContentTypeExportInterface, PreResolvableContentTypeInterface, ReferenceContentTypeInterface, PropertyMetadataMapperInterface
+class BlockContentType extends ComplexContentType implements ContentTypeExportInterface, PreResolvableContentTypeInterface, ReferenceContentTypeInterface
 {
     /**
      * @param BlockVisitorInterface[] $blockVisitors
      */
     public function __construct(
         private ContentTypeManagerInterface $contentTypeManager,
-        private SchemaMetadataProvider $schemaMetadataProvider,
         private iterable $blockVisitors,
     ) {
     }
@@ -466,41 +454,5 @@ class BlockContentType extends ComplexContentType implements ContentTypeExportIn
                 $child->setValue($oldValue);
             }
         }
-    }
-
-    public function mapPropertyMetadata(FieldMetadata $fieldMetadata): PropertyMetadata
-    {
-        $blockTypeSchemas = [];
-        foreach ($fieldMetadata->getTypes() as $blockType) {
-            $tag = $blockType->findTag('sulu.global_block');
-            if ($tag instanceof TagMetadata) {
-                $blockName = $tag->getAttributes()['global_block'] ?? null;
-                \assert(\is_string($blockName), 'Global block name is expected to be defined and a string.');
-
-                $blockTypeSchemas[] = new IfThenElseMetadata(
-                    new SchemaMetadata([
-                        new PropertyMetadata('type', true, new ConstMetadata($blockType->getKey())),
-                    ]),
-                    new RefSchemaMetadata('#/definitions/' . $blockName)
-                );
-
-                continue;
-            }
-
-            $blockTypeSchemas[] = new IfThenElseMetadata(
-                new SchemaMetadata([
-                    new PropertyMetadata('type', true, new ConstMetadata($blockType->getKey())),
-                ]),
-                $this->schemaMetadataProvider->getMetadata($blockType->getItems()),
-            );
-        }
-
-        return new PropertyMetadata(
-            $fieldMetadata->getName(),
-            $fieldMetadata->isRequired(),
-            new ArrayMetadata(
-                new AllOfsMetadata($blockTypeSchemas)
-            )
-        );
     }
 }
